+++ components/forms/dashboard/user-story.tsx
@@ -1,16 +1,16 @@
 import { FC, useEffect } from 'react';
 import { useForm } from 'react-hook-form';
 import { useDispatch } from 'react-redux';
-import { Button, Divider, Form, FormItem, Paragraph, TextArea, Title } from '../UI';
+import { Button, Divider, Form, FormItem, Paragraph, TextArea, Title } from '../../UI';
 import {
   createOrUpdateUserStory,
   getUserStory,
   setUserStory,
-} from '../../redux/redusers/forms/user-story';
-import { IStory } from '../../interfaces/story';
-import { useAppSelector } from '../../redux/store';
-import Alert from '../alert';
-import { storage } from '../../services/storage';
+} from '../../../redux/redusers/forms/user-story';
+import { IStory } from '../../../interfaces/story';
+import { useAppSelector } from '../../../redux/store';
+import Alert from '../../alert';
+import { storage } from '../../../services/storage';
 
 export const UserStoryForm: FC = () => {
   const { register, handleSubmit, setValue } = useForm();
